+++ WeatherCalendar/Location.swift
@@ -1,0 +1,15 @@
+//
+//  Location.swift
+//  WeatherCalendar
+//
+//  Created by Jaewon on 2022/05/12.
+//
+
+import Foundation
+
+enum Location {
+    struct Asia {
+        static let Seoul = (lat: 37.5683, lon: 126.9778)
+    }
+    
+}
